feature: rename package
