+++ src/main/java/decisionmakertool/metrics/templateimpl/CircularityErrors.java
@@ -1,4 +1,4 @@
-package decisionmakertool.metrics.templateImpl;
+package decisionmakertool.metrics.templateimpl;
 
 import ionelvirgilpop.drontoapi.except.UnexpectedErrorException;
 import ionelvirgilpop.drontoapi.pitfallmanager.Pitfall;
@@ -8,16 +8,17 @@ import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-public class IncompletenessErrors extends SmellErrorTemplate {
-    private static final  Integer []arrayPitfallsIncompleteness = {4,8,11,13};
+public class CircularityErrors extends SmellErrorTemplate{
+    private static final Integer []arrayPitfallsCircularity = {6};
 
-    public List<Pitfall> getListSmellErrors(String path){
+    public  List<Pitfall> getListSmellErrors(String path){
         List<Pitfall> listResult = new ArrayList<>();
         try {
-            SmellErrorTemplate.loadPitfallSelector(path, listResult, arrayPitfallsIncompleteness);
+            SmellErrorTemplate.loadPitfallSelector(path, listResult, arrayPitfallsCircularity);
         } catch (UnexpectedErrorException | IOException ex) {
-            Logger.getLogger(IncompletenessErrors.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(CircularityErrors.class.getName()).log(Level.SEVERE, null, ex);
         }
         return listResult;
     }
+
 }
